Revert "chore(ui): redirect to auth page when unauthenticated (#278)"
This reverts commit 432002f19d041e371170d886d1124ae6bb4c2c7b.
