+++ ui/src/lib/features/api-auth/store.ts
@@ -3,4 +3,4 @@
 import { writable } from 'svelte/store'
 
 export const authenticated = writable(false)
-export const apiAuthEnabled = writable<null | boolean>(null)
+export const apiAuthEnabled = writable(false)
