+++ ui/src/lib/utils/helpers.ts
@@ -6,7 +6,7 @@ import { get } from 'svelte/store'
 
 export const stringToSnakeCase = (name: string) => name.split(' ').join('-').toLocaleLowerCase()
 
-export async function fetchAPIAuthStatus(): Promise<Record<string, string>> {
+export async function fetchConfig(): Promise<Record<string, string>> {
   const response = await fetch('/auth-status')
   if (!response.ok) {
     throw new Error(`HTTP error! status: ${response.status}`)
@@ -24,14 +24,4 @@ export function createEventSource(path: string): EventSource {
     return new EventSource(url)
   }
   return new EventSource(path)
-}
-
-// Can't live in api-auth.ts because of sessionStorage usage in file, preventing this function
-// from being used in load functions
-export async function updateApiAuthEnabled() {
-  if (get(apiAuthEnabled) == null) {
-    const envVars = await fetchAPIAuthStatus()
-    // API Auth is only disabled when API_AUTH_DISABLED is set to 'true'
-    apiAuthEnabled.set(envVars.API_AUTH_DISABLED?.toLowerCase() !== 'true')
-  }
 }
