+++ ui/src/routes/+layout.svelte
@@ -5,20 +5,18 @@
   import 'flowbite'
   import { initFlowbite } from 'flowbite'
   import { onMount } from 'svelte'
+
   import { afterNavigate } from '$app/navigation'
   import { isSidebarExpanded, Navbar, Sidebar } from '$features/navigation'
   import { ToastPanel } from '$features/toast'
   import '../app.postcss'
   import { authenticated } from '$lib/features/api-auth/store'
   import { apiAuthEnabled } from '$lib/features/api-auth/store'
-  import Unauthenticated from '$components/Auth/component.svelte'
 
-  let path = ''
   // These initiFlowbite calls help load the js necessary to target components which use flowbite js
   // i.e. data-dropdown-toggle
   onMount(() => {
     initFlowbite()
-    path = window.location.pathname
   })
 
   afterNavigate(initFlowbite)
@@ -38,10 +36,6 @@
 >
   <div class="flex-grow overflow-hidden p-4 pt-6">
     <ToastPanel />
-    {#if $apiAuthEnabled && !$authenticated && path !== '/auth'}
-      <Unauthenticated />
-    {:else}
-      <slot />
-    {/if}
+    <slot />
   </div>
 </main>
